requirements.txt
audio_recorder_streamlit==0.0.8
langchain==0.1.6
langchain_community==0.0.19
langchain_openai==0.0.5
openai==1.12.0
pinecone_client==2.2.4
python-dotenv==1.0.1
streamlit==1.31.0
streamlit_chat==0.1.1
streamlit_float==0.3.2
tiktoken